requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
bcrypt
psycopg2-binary
Authlib
python-dotenv
requests
gunicorn